+++ NEWS.md
@@ -1,5 +1,6 @@
 # labelled (development version)
 
+* an additional argument `keep_var_label` for `remove_labels()`
 * bug fix for `unlabelled()` when applied on a vector
 * when using `unclass = TRUE` with `to_factor()`, attributes 
   are not removed anymore